Extract socket handler from `matchingController.js` and rename to `queueController.js`
